Ignore protocol on links
This page doesn't work on HTTPS, which is also supported by Github Pages. I use the "HTTPS Everywhere" browser extension, and navigating to this page didn't work because the non-secure links aren't being followed. I removed the protocol from those links, so the links will default to the protocol used when navigating to this page (either HTTP or HTTPS).